+++ vectors/cryptography_vectors/__about__.py
@@ -6,4 +6,4 @@ __all__ = [
     "__version__",
 ]
 
-__version__ = "44.0.0"
+__version__ = "45.0.0.dev1"