+++ CHANGELOG.rst
@@ -27,9 +27,9 @@ Changelog
   when using OpenSSL 3.2.0+.
 * Added support for the :class:`~cryptography.x509.Admissions` certificate extension.
 * Added basic support for PKCS7 decryption (including S/MIME 3.2) via
-  :class:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_der`, 
-  :class:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_pem`, and
-  :class:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_smime`.
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_der`, 
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_pem`, and
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_smime`.
 
 .. _v43-0-3: